My Blockbench Custom Model Data Pack
This is a texture pack that I have created some of the custom items using the Custom Model Data. For this example, I used an axe. I generated the model by imputting this command in the chat: /give @p minecraft:diamond_axe{CustomModelData:4}. The custom model data is 4 because it is the name in the json file and in the diamond_axe.json file, the overrides has the Custom Model Data set to 4.

Hope this helps!

< <DenisticCode>